+++ e2e/cypress/integration/enterprise/guest_accounts/guest_identification_spec.js
@@ -7,7 +7,7 @@
 // - Use element ID when selecting an element. Create one if none.
 // ***************************************************************
 
-// Group: @guest_account
+// Group: @enterprise @guest_account
 
 /**
  * Note: This test requires Enterprise license to be uploaded
@@ -22,8 +22,7 @@ describe('Guest Accounts', () => {
     let sysadmin;
     let testTeam;
     let testChannel;
-    let adminMFASecret = '';
-    let guestUser;
+    let adminMFASecret;
     const username = 'g' + getRandomId(); // username has to start with a letter.
 
     before(() => {
@@ -38,13 +37,9 @@ describe('Guest Accounts', () => {
         cy.apiAdminLogin().then((res) => {
             sysadmin = res.user;
         });
-
-        cy.apiCreateGuestUser().then(({guest}) => {
-            guestUser = guest;
-        });
     });
 
-    afterEach(() => {
+    after(() => {
         // # Login back as admin.
         const token = authenticator.generateToken(adminMFASecret);
         cy.apiAdminLoginWithMFA(token);
@@ -189,74 +184,4 @@ describe('Guest Accounts', () => {
         });
         cy.apiLogout();
     });
-
-    it('MM-T1411 Update Guest Users in User Management when Guest feature is disabled', () => {
-        // # Navigate to Guest Access page.
-        cy.visit('/admin_console/authentication/guest_access');
-
-        // # Enable guest accounts.
-        cy.findByTestId('GuestAccountsSettings.Enabletrue').check();
-
-        // # Click "Save".
-        cy.get('#saveSetting').then((btn) => {
-            if (btn.is(':enabled')) {
-                btn.on('click');
-
-                cy.waitUntil(() => cy.get('#saveSetting').then((el) => {
-                    return el[0].innerText === 'Save';
-                }));
-            }
-        });
-
-        // # Go to System Console ➜ Users.
-        cy.visit('/admin_console/user_management/users');
-
-        // # Ensure there are active Guest users.
-        checkUserListStatus(guestUser, 'Guest');
-
-        // # Navigate to System Console ➜ Guest Access.
-        cy.visit('/admin_console/authentication/guest_access');
-
-        // # Set Enable Guest Access to false.
-        cy.findByTestId('GuestAccountsSettings.Enablefalse').check();
-
-        // # Click "Save".
-        cy.get('#saveSetting').scrollIntoView().click();
-        cy.get('#confirmModal').should('be.visible').within(() => {
-            cy.get('#confirmModalButton').should('have.text', 'Save and Disable Guest Access').click();
-        });
-
-        // # Go to System Console ➜ Users.
-        cy.visit('/admin_console/user_management/users');
-
-        // * Guest users are shown as "Inactive".
-        checkUserListStatus(guestUser, 'Inactive');
-
-        // # Navigate to Guest Access page.
-        cy.visit('/admin_console/authentication/guest_access');
-
-        // # Enable guest accounts.
-        cy.findByTestId('GuestAccountsSettings.Enabletrue').check();
-
-        // # Click "Save".
-        cy.get('#saveSetting').scrollIntoView().click();
-
-        // # Go to System Console ➜ Users.
-        cy.visit('/admin_console/user_management/users');
-
-        // * Guest users are shown as "Inactive".
-        checkUserListStatus(guestUser, 'Inactive');
-    });
-
-    function getInnerText(el) {
-        return el[0].innerText.replace(/\n/g, '').replace(/\s/g, ' ');
-    }
-
-    function checkUserListStatus(user, status) {
-        cy.get('#searchUsers').should('be.visible').type(user.username);
-        cy.get('#selectUserStatus').select(status);
-        cy.get('.more-modal__details > .more-modal__name').should('be.visible').then((el) => {
-            expect(getInnerText(el)).contains(`@${user.username}`);
-        });
-    }
 });
